cfg: change authors to `the KaA group`
Displaying only three members in the author node is not representative
nor fair of/to the rest of the group, and adding the names of all 5
members (and possibly more as time passes) introduces unnecessary bloat
to the configuration files.

Going forward, all `author` nodes will be populated with `the Kerbae ad
Astra group` instead of any individual('s|s') name(s).